Moved is_float() to utilities
